move debug down a level
